Refactored all Entity classes into their own package
