+++ core/src/main/java/core/game/entities/BaseMonster.java
@@ -1,4 +1,4 @@
-package core.game.logic;
+package core.game.entities;
 
 public abstract class BaseMonster extends Entity {
