+++ core/src/main/java/core/gdx/wad/MyGDxTest.java
@@ -2,18 +2,13 @@ package core.gdx.wad;
 
 import com.badlogic.gdx.Game;
 import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.utils.Array;
-import core.game.logic.Entity;
 import core.game.logic.GameLogic;
-import core.game.logic.PlayerPawn;
 import core.level.info.LevelData;
 import core.wad.funcs.WadFuncs;
 import net.mtrop.doom.WadFile;
 
 import java.io.IOException;
-
-import static com.badlogic.gdx.Gdx.audio;
 
 /** {@link com.badlogic.gdx.ApplicationListener} implementation shared by all platforms. */
 public class MyGDxTest extends Game {
